refactor: Move criteria card to its own folder
