+++ src/components/Rubric/ObjectiveRow.tsx
@@ -1,8 +1,8 @@
 import { AddCriteriaDialog } from "@/screens/edit-rubric/dialogs/add-criteria/AddCriteriaDialog";
 import { Objective } from "@/types/entities/rubric";
 
-import { CriteriaCard } from "./CriteriaCard";
-import { ObjectiveCard } from "./ObjectiveCard";
+import { CriteriaCard } from "./CriteriaCard/CriteriaCard";
+import { ObjectiveCard } from "./ObjectiveCard/ObjectiveCard";
 
 interface ObjectiveRowProps {
   objective: Objective;
